requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
mssql-django>=1.4.0
pyodbc>=5.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
django-apscheduler>=0.6.2
Pillow>=10.0.0
google-genai>=0.2.0
# Fallback to deprecated package if new one not available
google-generativeai>=0.3.0; python_version < "3.11"
# Bloomerp Framework
git+https://github.com/DavidBloomer11/Bloomerp.git@main
django-htmx>=1.18.0
# Bloomerp Dependencies
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0

